add changed approach to lab
This approach involves the same strategy of splitting the pattern into two halves, however this one does not have a center, rather the center element (if there is one) will be placed into the right side of the halves. We will compare them seeing if there are differences, when there are we will check the right neighbor (n+1) and determine if removing that element (n) will be replaced by the correct element. More testing should be done, but this covers some cases so far.